// webui/src/components/panel/hashing/lists/__generated__/ListsTableFragment.graphql.ts
export type ListsTableFragment$data = {
  readonly lists: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly tags: ReadonlyArray<string> | null;
        readonly creator: {
          readonly id: string;
          readonly username: string;
          readonly matrixLinks: ReadonlyArray<string> | null;
        };
      };
    }>;
  } | null;
  readonly " $fragmentType": "ListsTableFragment";
};
export type ListsTableFragment$key = {
  readonly " $data"?: ListsTableFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ListsTableFragment">;
};